+++ hadoop/src/test/java/org/apache/metamodel/util/HdfsResourceTest.java
@@ -28,18 +28,33 @@ public class HdfsResourceTest extends TestCase {
         final HdfsResource res1 = new HdfsResource("hdfs://localhost:9000/home/metamodel.txt");
         assertEquals("hdfs://localhost:9000/home/metamodel.txt", res1.getQualifiedPath());
         assertEquals("metamodel.txt", res1.getName());
+        assertEquals("hdfs", res1.getScheme());
 
         final HdfsResource res2 = new HdfsResource("localhost", 9000, "/home/metamodel.txt");
         assertEquals("hdfs://localhost:9000/home/metamodel.txt", res2.getQualifiedPath());
         assertEquals("metamodel.txt", res2.getName());
+        assertEquals("hdfs", res2.getScheme());
 
         assertEquals(res1, res2);
 
         final HdfsResource res3 = new HdfsResource("localhost", 9000, "/home/apache.txt");
         assertEquals("hdfs://localhost:9000/home/apache.txt", res3.getQualifiedPath());
         assertEquals("apache.txt", res3.getName());
+        assertEquals("hdfs", res3.getScheme());
 
         assertFalse(res3.equals(res1));
+        
+        final HdfsResource res4 = new HdfsResource("emrfs://localhost:9000/home/metamodel.txt");
+        assertEquals("emrfs://localhost:9000/home/metamodel.txt", res4.getQualifiedPath());
+        assertEquals("metamodel.txt", res4.getName());
+        assertEquals("emrfs", res4.getScheme());
+        
+        assertFalse(res4.equals(res1));
+    }
+    
+    public void testNoHostOrPortInUrl() throws Exception {
+        final HdfsResource resource = new HdfsResource("emrfs:///home/metamodel.txt");
+        assertEquals("emrfs:///home/metamodel.txt", resource.getQualifiedPath());
     }
     
     public void testSerialization() throws Exception {